+++ src/games/super-mario-land/objects/box.ts
@@ -23,13 +23,35 @@ export class Box extends Phaser.GameObjects.Sprite {
 
   private initSprite() {
     // sprite
-    this.setOrigin(0.5, 0.5);
+    this.setOrigin(0, 0);
+    this.setFrame(0);
 
     // physics
     this.currentScene.physics.world.enable(this);
+    this.body.setSize(8, 8);
     this.body.setAllowGravity(false);
     this.body.setImmovable(true);
   }
 
-  update(): void {}
+  update(): void {
+    if (this.body.touching.down && this.active) {
+      // OPTIMIZE TWEEN WITH FUNCTION TO SET ACTIVE TO FALSE AND FRAME TO 1
+      this.currentScene.tweens.add({
+        targets: this,
+        props: { y: this.y - 10 },
+        delay: 0,
+        duration: 60,
+        ease: "Power0",
+        easeParams: null,
+        hold: 0,
+        repeat: 0,
+        repeatDelay: 0,
+        yoyo: true,
+        paused: false
+      });
+
+      this.active = false;
+      this.setFrame(1);
+    }
+  }
 }