+++ src/games/super-mario-land/scenes/menu-scene.ts
@@ -46,7 +46,9 @@ export class MenuScene extends Phaser.Scene {
 
   update(): void {
     if (this.startKey.isDown) {
+      this.scene.start("HUDScene");
       this.scene.start("GameScene");
+      this.scene.bringToTop("HUDScene");
     }
   }
 }